+++ Nexus.Hub/Nexus.WeightTracker.Api/Domain/ErrorHandling/DomainErrorProvider.cs
@@ -4,7 +4,7 @@ namespace Nexus.WeightTracker.Api.Domain.ErrorHandling;
 
 public static class DomainErrorProvider
 {
-    public static Error InvalidRecordedValue(decimal recordedValue)
+    public static Error InvalidRecordedValue(double recordedValue)
     {
         return new InvalidRecordedValue(recordedValue);
     }
